feat(*): pull out runtime into a tiny crate
