+++ lib/dal/BUCK
@@ -17,8 +17,8 @@ rust_library(
         "//lib/si-hash:si-hash",
         "//lib/si-layer-cache:si-layer-cache",
         "//lib/si-pkg:si-pkg",
+        "//lib/si-runtime-rs:si-runtime",
         "//lib/si-settings:si-settings",
-        "//lib/si-service:si-service",
         "//lib/si-std:si-std",
         "//lib/telemetry-rs:telemetry",
         "//lib/telemetry-nats-rs:telemetry-nats",
