+++ lib/si-runtime-rs/BUCK
@@ -1,0 +1,9 @@
+load("@prelude-si//:macros.bzl", "rust_library")
+
+rust_library(
+    name = "si-runtime",
+    deps = [
+        "//third-party/rust:tokio",
+    ],
+    srcs = glob(["src/**/*.rs"]),
+)
