+++ lib/si-service/BUCK
@@ -3,6 +3,7 @@ load("@prelude-si//:macros.bzl", "rust_library")
 rust_library(
     name = "si-service",
     deps = [
+        "//lib/si-runtime-rs:si-runtime",
         "//lib/telemetry-application-rs:telemetry-application",
         "//lib/telemetry-rs:telemetry",
         "//third-party/rust:color-eyre",
